Fix: Shared 페이지 구현 중(DEFAULT USER ID 사용 중)
